DEV: Refactor-img_proc to image_processing, mathops to arithmetic
In an effort to simplify the user experience for image processing,
image processing folders and modules are being refactored to more
clearly state what kind of tools each module contains. This
initial refactor specifically changes the folder name:
skxray/img_proc to skxray/image_processing and the module mathops.py to
arithmetic.py. This also addresses a long standing question from Tom
about why I was beholden to the name mathops (which stood for math
operations). I'm hoping that these changes add an additional level of
clarity.
