+++ skxray/image_processing/arithmetic.py
