+++ skxray/tests/test_io/test_netCDF_io.py
@@ -32,7 +32,7 @@ def test_net_cdf_io():
     #TODO: The reader functions are complete. Writer functions still need to be
     #   Added to the am-IO function set. As soon as that is complete a series of
     #   read --> write --> read test functions will be added to this file.
-    #TODO: this will be address after addition of img_proc functions to vistrails
+    #TODO: this will be address after addition of image_processing functions to vistrails
     #data, md_dict = ncd.load_netCDF(test_data)
     #eq_(md_dict['operator'], 'Iltis')
     #eq_(data.shape, (470, 695, 695))
